Refactor shared utils into focused modules with shim
